Network layer refactoring - add RequestSender. Fixed all unit tests.
