+++ AndroidSDKCore/src/main/java/com/leanplum/internal/Log.java
@@ -188,7 +188,8 @@ public class Log {
       HashMap<String, Object> params = new HashMap<>();
       params.put(Constants.Params.TYPE, Constants.Values.SDK_LOG);
       params.put(Constants.Params.MESSAGE, message);
-      RequestOld.post(Constants.Methods.LOG, params).sendEventually();
+      RequestOld request = RequestOld.post(Constants.Methods.LOG, params);
+      RequestSender.getInstance().sendEventually(request);
     } catch (Throwable t) {
       android.util.Log.e("Leanplum", "Unable to send log.", t);
     } finally {
