+++ AndroidSDKTests/src/test/java/com/leanplum/_whitebox/utilities/RequestHelper.java
@@ -20,14 +20,8 @@
  */
 package com.leanplum._whitebox.utilities;
 
-import android.util.Log;
-
 import com.leanplum.internal.RequestOld;
 
-import org.json.JSONArray;
-import org.json.JSONObject;
-
-import java.lang.reflect.Field;
 import java.util.Map;
 
 /**
@@ -54,40 +48,6 @@ public class RequestHelper extends RequestOld {
    */
   public static void addRequestHandler(RequestHandler handler) {
     sRequestHandler = handler;
-  }
-
-  @Override
-  public void sendEventually() {
-    // workaround to send request now and not save it into prefs for later use
-    try {
-      Field sentField = RequestOld.class.getDeclaredField("sent");
-      sentField.setAccessible(true);
-
-      boolean sent = sentField.getBoolean(this);
-
-      if (!sent) {
-        super.sendEventually();
-        super.sendIfConnected();
-      }
-
-    } catch (Exception e) {
-      Log.e(RequestHelper.class.getSimpleName(), "Could not access private field \"sent\"");
-    }
-  }
-
-  @Override
-  protected void parseResponseBody(JSONObject responseBody, Exception error) {
-    try {
-      // attach the uuid we generated
-      JSONArray jsonArray = responseBody.getJSONArray("response");
-      for (int i = 0; i < jsonArray.length(); i++) {
-        JSONObject jsonObject = jsonArray.getJSONObject(i);
-        jsonObject.put("reqId", requestId());
-      }
-    } catch (Exception e) {
-      // ignore
-    }
-    super.parseResponseBody(responseBody, error);
   }
 
   /**
